test: consolidate settings tests at page level
